+++ src/lox/parser.rs
@@ -85,7 +85,7 @@ impl Parser {
 
         let body = self.block_stmt()?;
 
-        Ok(FunStmt::new(name, params, body).into())
+        Ok(FunStmt::new(name, params, body, None).into())
     }
 
     fn var_dec(&mut self) -> Result<Stmt, Err> {